extract ConstantPlaceholder into a module;
